% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_negative_controls.R
\name{add_negative_controls}
\alias{add_negative_controls}
\title{add_negative_controls}
\usage{
add_negative_controls(
  bbama_model,
  data,
  id = bbama_model$id,
  likelihood = bbama_model$likelihood,
  fixed = NULL,
  y1 = NULL,
  y0 = NULL,
  pt1 = NULL,
  pt0 = NULL,
  est = NULL,
  se = NULL,
  same_args = TRUE
)
}
\description{
add_negative_controls
}
