% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_likelihood.R
\name{set_likelihood}
\alias{set_likelihood}
\title{set_likelihood}
\usage{
set_likelihood(
  bbama_model,
  likelihood = c("poisson", "normal", "grid"),
  fixed = NULL,
  y1 = NULL,
  y0 = NULL,
  pt1 = NULL,
  pt0 = NULL,
  est = NULL,
  se = NULL,
  ll = NULL,
  point = NULL,
  value = NULL,
  n_vals = NULL,
  min_val = NULL,
  max_val = NULL,
  extreme_min_val = log(1/100),
  extreme_max_val = log(100),
  extrapolate_formula = value ~ poly(point, 4),
  sep = ";",
  ...
)
}
\description{
set_likelihood
}
