% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  n_outcomes = 100,
  databases = c("A", "B", "C", "D", "E"),
  x1_databases = c(),
  x2_databases = c(),
  n_per_database = round(1e+05 * runif(5)),
  min_rate = 0,
  max_rate = 0.1,
  obs_days = 30,
  prop_included = 1,
  mu = log(1.5),
  tau = 0.1,
  lambda = log(1.5),
  eta = 0.1,
  min_gamma = 0.1,
  max_gamma = 0.3,
  x1_fixed = 0,
  x2_fixed = 0,
  include_ll = TRUE
)
}
\description{
simulate_data
}
